requirements.txt
python-pptx>=0.6.23
Pillow>=10.0.0
reportlab>=4.0.0
